Removed higher half kernel functionality
In this new VM there will be no need for a higher half kernel because
the VM will use the entire address space. All code runs in ring0, the
difference is the kernel only lets audited code run. The idea is that
code gets compiled to an IR and that IR is then translated by the VM to
machine code. The IR will not allow pointers or other potentially unsafe
actions, but still be as flexible as possible and expose as much of the
machine as possible.

We'll see how far this goes. It would be nice to fully expose a good
deal of the hardware above the VM level so most things could be written
in a VM targeted language.