Trying to allow optional makefile command line args
The kernel has a couple of optional debug/test flags that can be defined
while compiling, before they would override the required flags. Now they
are appended after the required flags. This may not be the proper way to
do it, but it works for now and will be investigated further later.