Making progress on the APIC.
Now the APIC is enabled and appears to work. To test, interrupts have to
be enabled by added an __asm__("sti") after apic_init() is called. Will
be added later.

Currently the APIC is being used in compatability mode with the old 8259
PIC. The next work will be to find and enable I/O APICs instead of the
8259.