Removed mkterm.sh and added it to the makefile
I suppose there was no need for an external script. I was playing with
different methods before and gnumake wasn't playing along so I moved to
an external script. The solution is much simpler than what it was an
gnumake seems to be playing along, so moved it back.

There does appear to be a race condition, but I'm not sure how to fix
it, and it doesn't seem to happen very often.