The paging module now supports more than 1GiB of RAM
Before many assumptions were being made about where the memory regions
were and what was available for us. When testing on real hardware the
'choose largest region' broke because that region was not directly after
the kernel and when trying to map the region things broke.

Also the paging_init() function was using the kmap_page_range() function
which relies on the physical allocator. Unfortunately the physical
allocator is not setup at this point and that created a fault/double
fault. Now it manually checks and allocates the tables, this is a one
time setup anyway, after this the physical memory allocator can be used
without issue.