Interrupts work! Yay!
With the addition of the previous mmap fixes interrupts even work on
real hardware! It's always a good thing to incrementally test things on
hardware because QEMU is a little too forgiving in some areas. For
example it allowed the use of unmapped pages without faulting,
unfortunately real hardware was not so nice :)