Found the problem with interrupts.
The ioapic_read/write functions were using ioapic_base_virt_address,
which was no longer needed with the identity mapping. Instead it has
been reassigned to the base address of the first I/O APIC, this should
be okay, because the I/O APIC addresses are verified to be sequential.